+++ test/GovNFTFactory.t.sol
@@ -63,8 +63,8 @@ contract GovNFTFactoryTest is BaseTest {
             testToken,
             address(recipient),
             TOKEN_100K,
-            block.timestamp,
-            block.timestamp + WEEK * 2,
+            uint40(block.timestamp),
+            uint40(block.timestamp) + WEEK * 2,
             WEEK
         );
         notAdmin.approve(testToken, address(_govNFT), TOKEN_100K);
@@ -73,8 +73,8 @@ contract GovNFTFactoryTest is BaseTest {
             testToken,
             address(recipient),
             TOKEN_100K,
-            block.timestamp,
-            block.timestamp + WEEK * 2,
+            uint40(block.timestamp),
+            uint40(block.timestamp) + WEEK * 2,
             WEEK
         );
     }
@@ -90,8 +90,8 @@ contract GovNFTFactoryTest is BaseTest {
             testToken,
             address(recipient),
             TOKEN_100K,
-            block.timestamp,
-            block.timestamp + WEEK * 2,
+            uint40(block.timestamp),
+            uint40(block.timestamp) + WEEK * 2,
             WEEK
         );
         notAdmin.approve(testToken, address(_govNFT), TOKEN_100K);
@@ -101,8 +101,8 @@ contract GovNFTFactoryTest is BaseTest {
             testToken,
             address(recipient),
             TOKEN_100K,
-            block.timestamp,
-            block.timestamp + WEEK * 2,
+            uint40(block.timestamp),
+            uint40(block.timestamp) + WEEK * 2,
             WEEK
         );
     }
